requirements.txt
-ip==20.1.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pip==20.2.2
setuptools==47.1.0
yarl==1.5.1